Work_Project_Api/requirement.txt
allure-pytest==2.15.2
allure-python-commons==2.15.2
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dotenv==0.9.9
greenlet==3.3.0
idna==3.11
iniconfig==2.1.0
mimesis==18.0.0
packaging==25.0
pip==25.3
playwright==1.57.0
pluggy==1.6.0
pydantic==2.12.5
pydantic-core==2.41.5
pyee==13.0.0
pygments==2.19.2
pytest==9.0.2
pytest-sugar==1.1.1
python-dotenv==1.2.1
requests==2.32.5
termcolor==3.2.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
